# src/main/resources/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "接口说明文档",
    "version" : "v1",
    "title" : "接口说明文档"
  },
  "host" : "localhost:8080",
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/chart/show/{foodid}" : {
      "get" : {
        "summary" : "使用量图表数据获取",
        "description" : "",
        "operationId" : "storeChartShow",
        "parameters" : [ {
          "name" : "type",
          "in" : "query",
          "description" : "图表种类",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "foodid",
          "in" : "path",
          "description" : "菜品编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
...
    "/store/ask" : {
      "put" : {
        "summary" : "菜品购入请求",
        "description" : "",
        "operationId" : "askedStore",
        "parameters" : [ {
          "name" : "foodid",
          "in" : "query",
          "description" : "菜品编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "changenum",
          "in" : "query",
          "description" : "请求量",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/store/chart/{id}" : {
      "get" : {
        "summary" : "打开图表显示",
        "description" : "",
        "operationId" : "storeChart",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "菜品编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Model"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/store/use" : {
      "put" : {
        "summary" : "菜品使用",
        "description" : "",
        "operationId" : "useFood",
        "parameters" : [ {
          "name" : "foodid",
          "in" : "query",
          "description" : "菜品编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "changenum",
          "in" : "query",
          "description" : "使用量",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "object"
              }
            }
          }
        }
      }
    },
    "/user" : {
      "post" : {
        "summary" : "添加账户",
        "description" : "",
        "operationId" : "addAccout",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "用户详细信息",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UserLogin"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RestMessage"
            }
          }
        }
      }
    },
    "/user/login" : {
      "post" : {
        "summary" : "判断用户登录",
        "description" : "",
        "operationId" : "testAjax",
        "parameters" : [ {
          "name" : "account",
          "in" : "query",
          "description" : "用户账号",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "用户密码",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RestMessage"
            }
          }
        }
      }
    },
    "/user/register" : {
      "post" : {
        "summary" : "新增供应商",
        "description" : "",
        "operationId" : "providerSignUp",
        "parameters" : [ {
          "name" : "account",
          "in" : "query",
          "description" : "供应商账号",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "供应商密码",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "username",
          "in" : "query",
          "description" : "供应商用户名",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RestMessage"
            }
          }
        }
      }
    },
    "/user/{id}" : {
      "delete" : {
        "summary" : "删除账户",
        "description" : "",
        "operationId" : "deleteAccout",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "删除账号的编号",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RestMessage"
            }
          }
        }
      }
    },
    "/users" : {
      "get" : {
        "summary" : "显示所有用户",
        "description" : "",
        "operationId" : "accountMsg",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/HttpServletRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/RestMessage"
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "BufferedReader" : {
      "type" : "object"
    },
    "Cookie" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "comment" : {
          "type" : "string"
        },
        "domain" : {
          "type" : "string"
        },
        "maxAge" : {
          "type" : "integer",
          "format" : "int32"
        },
        "path" : {
          "type" : "string"
        },
        "secure" : {
          "type" : "boolean",
          "default" : false
        },
        "version" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Enumeration" : {
      "type" : "object"
    },
    "HttpServletRequest" : {
      "type" : "object",
      "properties" : {
        "method" : {
          "type" : "string"
        },
        "requestedSessionIdValid" : {
          "type" : "boolean",
          "default" : false
        },
        "pathTranslated" : {
          "type" : "string"
        },
        "requestedSessionIdFromCookie" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionIdFromURL" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionIdFromUrl" : {
          "type" : "boolean",
          "default" : false
        },
        "requestedSessionId" : {
          "type" : "string"
        },
        "session" : {
          "$ref" : "#/definitions/HttpSession"
        },
        "userPrincipal" : {
          "$ref" : "#/definitions/Principal"
        },
        "contextPath" : {
          "type" : "string"
        },
        "requestURL" : {
          "$ref" : "#/definitions/StringBuffer"
        },
        "headerNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "remoteUser" : {
          "type" : "string"
        },
        "authType" : {
          "type" : "string"
        },
        "pathInfo" : {
          "type" : "string"
        },
        "servletPath" : {
          "type" : "string"
        },
        "requestURI" : {
          "type" : "string"
        },
        "queryString" : {
          "type" : "string"
        },
        "cookies" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Cookie"
          }
        },
        "scheme" : {
          "type" : "string"
        },
        "inputStream" : {
          "$ref" : "#/definitions/ServletInputStream"
        },
        "protocol" : {
          "type" : "string"
        },
        "reader" : {
          "$ref" : "#/definitions/BufferedReader"
        },
        "contentLength" : {
          "type" : "integer",
          "format" : "int32"
        },
        "contentType" : {
          "type" : "string"
        },
        "locale" : {
          "$ref" : "#/definitions/Locale"
        },
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "characterEncoding" : {
          "type" : "string"
        },
        "parameterMap" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        },
        "parameterNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "localName" : {
          "type" : "string"
        },
        "remoteHost" : {
          "type" : "string"
        },
        "remotePort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "serverName" : {
          "type" : "string"
        },
        "localPort" : {
          "type" : "integer",
          "format" : "int32"
        },
        "locales" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "remoteAddr" : {
          "type" : "string"
        },
        "secure" : {
          "type" : "boolean",
          "default" : false
        },
        "localAddr" : {
          "type" : "string"
        },
        "serverPort" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "HttpSession" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "creationTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletContext" : {
          "$ref" : "#/definitions/ServletContext"
        },
        "maxInactiveInterval" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sessionContext" : {
          "$ref" : "#/definitions/HttpSessionContext"
        },
        "lastAccessedTime" : {
          "type" : "integer",
          "format" : "int64"
        },
        "valueNames" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "new" : {
          "type" : "boolean",
          "default" : false
        }
      }
    },
    "HttpSessionContext" : {
      "type" : "object",
      "properties" : {
        "ids" : {
          "$ref" : "#/definitions/Enumeration"
        }
      }
    },
    "Locale" : {
      "type" : "object",
      "properties" : {
        "language" : {
          "type" : "string"
        },
        "country" : {
          "type" : "string"
        },
        "displayCountry" : {
          "type" : "string"
        },
        "displayLanguage" : {
          "type" : "string"
        },
        "displayName" : {
          "type" : "string"
        },
        "displayScript" : {
          "type" : "string"
        },
        "displayVariant" : {
          "type" : "string"
        },
        "extensionKeys" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "iso3Country" : {
          "type" : "string"
        },
        "iso3Language" : {
          "type" : "string"
        },
        "script" : {
          "type" : "string"
        },
        "unicodeLocaleAttributes" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "unicodeLocaleKeys" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "variant" : {
          "type" : "string"
        }
      }
    },
    "Model" : {
      "type" : "object"
    },
...
    "Principal" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        }
      }
    },
...
    "RestMessage" : {
      "type" : "object",
      "properties" : {
        "count" : {
          "type" : "string",
          "description" : "数据数量"
        },
        "data" : {
          "type" : "object",
          "description" : "返回数据"
        },
        "code" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "返回编号"
        },
        "msg" : {
          "type" : "string",
          "description" : "返回信息"
        }
      },
      "description" : "返回信息"
    },
    "ServletContext" : {
      "type" : "object",
      "properties" : {
        "attributeNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "initParameterNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletContextName" : {
          "type" : "string"
        },
        "majorVersion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "minorVersion" : {
          "type" : "integer",
          "format" : "int32"
        },
        "contextPath" : {
          "type" : "string"
        },
        "servlets" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "servletNames" : {
          "$ref" : "#/definitions/Enumeration"
        },
        "serverInfo" : {
          "type" : "string"
        }
      }
    },
    "ServletInputStream" : {
      "type" : "object"
    },
    "StringBuffer" : {
      "type" : "object"
    },
    "UserLogin" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "account" : {
          "type" : "string"
        },
        "username" : {
          "type" : "string"
        },
        "password" : {
          "type" : "string"
        },
        "identity" : {
          "type" : "string"
        }
      }
    }
  }
}